fix(StrikeAndBallCount): bug that cause the argument to change
In the past, it requested to change the argument as it is.
So, if an immutable object is received as an argument, an exception is raised, which can cause problems due to shallow copying even if no exception is raised.
So, change the caller to copy and use the argument.

Before: 

public StrikeAndBallCount(List<Integer> target, List<Integer> guessValue) {
    digitNumber = target.size();
    setStrikeCountAndCheckIfStrike(copyTarget, copyGuessValue);
    setBallCount(copyTarget, copyGuessValue);
}


After:

public StrikeAndBallCount(List<Integer> target, List<Integer> guessValue) {
    List<Integer> copyTarget = new ArrayList<>(target);
    List<Integer> copyGuessValue = new ArrayList<>(guessValue);
    digitNumber = target.size();
    setStrikeCountAndCheckIfStrike(copyTarget, copyGuessValue);
    setBallCount(copyTarget, copyGuessValue);
}