fix: move startGame method in while loop
To restart a game "referee.startGame()" should be located
on the begining of a while loop in main method.

It is because the isGameInProgress variable in referee instance
is changed only when to call startGame().

After win a game isGameInProgress is 'false'. In order to restart a
game is GameInProgress should be changed to 'true'.