refactor<making Random> : int[] randomNumbers to instance variable, add initializer to it
CreateRandom class is in charge of making randomNumber array and pass it to Game object. In terms of stability, it would be better to hold random number array in instance, not locally.

Before:
int[] randomNumbers managed inside makeRandomNums() method, which was local vairable
After:
int[] randomNumbers is instance method, add initializer to clear content of randomNumbers