fix: change participant instances to be non-static
1. the instance variable "willPlay" in SportsCasters
("toTerminateGame" in Referee before) is being stored and can be tracked
when run Application.main repeatedly because it is in static memory.
So to solve this problem and prevent unoccured problems move all
participant instances in a method and make them non-static.

2. refactor: at the very before commit, some features in Referee
was distributed to other classes. These changes is applied to
the Application.java